% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_this.R
\name{scale_this}
\alias{scale_this}
\title{Scale a numeric vector without converting to a matrix}
\usage{
scale_this(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of scaled values.
}
\description{
This function scales and centres a numeric vector by subtracting the mean and dividing by the standard deviation.
Unlike `scale()`, it returns a numeric vector, not a matrix. Note this function does not allow control over centering or scaling.
}
\examples{
iris_dat <- head(iris)
scale_this(iris_dat$Sepal.Length)
scale_this(c(iris_dat$Sepal.Length, NA))
}
